requirements.txt
allure-pytest==2.9.45
Faker==13.3.4
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask==2.1.1
passlib==1.7.4
PyJWT==2.3.0
pytest==7.1.1
python-dotenv==0.20.0